server/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
python-dotenv>=1.0.0
torch>=2.0.0
silero-vad>=1.0.0
numpy>=1.24.0
httpx>=0.25.0
pydantic>=2.0.0
transformers>=4.40.0
onnxruntime>=1.17.0